Split view transition related css property parsing out into its own file https://bugs.webkit.org/show_bug.cgi?id=279298
Reviewed by Tim Nguyen and Darin Adler.

Just splits the consumers out into their own file. No additional cleanup performed.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+ViewTransition.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+ViewTransition.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/css/process-css-properties.py:

Canonical link: https://commits.webkit.org/283306@main
